producers/pyproject.toml
[project]
name = "activity-producer"
version = "0.1.0"
description = "Kafka producer for raw_activities topic with Avro serialization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "confluent-kafka[avro]>=2.3.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
produce-activities = "activity_producer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/activity_producer"]
